# Use repartition in window functions to speed up