@Jimexist 2021-06-21
benchmark run

```
Benchmarking window empty over, aggregate functions: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.9s, enable flat sampling, or reduce sample count to 60.
window empty over, aggregate functions
                        time:   [1.1699 ms 1.1755 ms 1.1811 ms]
                        change: [-37.629% -36.780% -36.002%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe

Benchmarking window empty over, built-in functions: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.3s, enable flat sampling, or reduce sample count to 60.
window empty over, built-in functions
                        time:   [1.0180 ms 1.0236 ms 1.0299 ms]
                        change: [-26.237% -25.544% -24.778%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe

window order by, aggregate functions
                        time:   [8.0163 ms 8.0741 ms 8.1409 ms]
                        change: [-3.3019% -1.9941% -0.6139%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe

window order by, built-in functions
                        time:   [7.4822 ms 7.5313 ms 7.5843 ms]
                        change: [-0.8959% +0.1569% +1.1590%] (p = 0.77 > 0.05)
                        No change in performance detected.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild

window partition by, u64_wide, aggregate functions
                        time:   [20.548 ms 20.664 ms 20.789 ms]
                        change: [+7.8065% +9.0739% +10.344%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

window partition by, u64_narrow, aggregate functions
                        time:   [7.2287 ms 7.2704 ms 7.3122 ms]
                        change: [+11.989% +13.130% +14.218%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild

window partition by, u64_wide, built-in functions
                        time:   [19.238 ms 19.360 ms 19.488 ms]
                        change: [+16.724% +18.053% +19.384%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

window partition by, u64_narrow, built-in functions
                        time:   [6.4687 ms 6.5140 ms 6.5591 ms]
                        change: [+8.1468% +9.2522% +10.283%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

window partition and order by, u64_wide, aggregate functions
                        time:   [27.575 ms 27.785 ms 28.013 ms]
                        change: [+9.4015% +10.538% +11.674%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high mild

Benchmarking window partition and order by, u64_narrow, aggregate functions: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.5s, or reduce sample count to 90.
window partition and order by, u64_narrow, aggregate functions
                        time:   [54.981 ms 55.324 ms 55.702 ms]
                        change: [-0.6417% +0.7593% +2.0563%] (p = 0.29 > 0.05)
                        No change in performance detected.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe

window partition and order by, u64_wide, built-in functions
                        time:   [25.673 ms 25.817 ms 25.969 ms]
                        change: [+7.6207% +8.7502% +9.8676%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

window partition and order by, u64_narrow, built-in functions
                        time:   [16.124 ms 16.211 ms 16.308 ms]
                        change: [+0.5055% +1.5834% +2.6646%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
```