@Dandandan 2021-06-22
Nice, almost 5 times improvement for some queries!

But do those have to do with the changes wrt sort or with regards to partition by?