@Jimexist 2021-06-23
> Nice, almost 5 times improvement for some queries!
> 
> But do those have to do with the changes wrt sort or with regards to partition by?

I don't expect that much of an improvement. let me redo the benchmark after #573 is merged.